Version 3.9.1 of ATBI Records Upload
This extract contains ALL observations that have been identified to
Genus, and DOES NOT filter out those that have less than 30
observations. In addition, there are many observations that fall outside
of the MAXENT model Area of Interest. The Maxent model will only use
species with 30 or more observations within the study area of interest.