Fix bug when returning to initial search
When using the breadcrumb to return to the initial search
tool, the "last page" of the search tool was not being updated
to reflect that the comparison page was left.  This caused a bug
when switching between the map options and the search tool, as
the "last page" the search page would attempt to go to was wrong.