Change data quality to model confidence
Per request of NPS, changed the "DATA QUALITY" terminology in the
color legend to "MODEL CONFIDENCE" to better reflect the meaning
of the content here.

To allow for this, I extended the width of
the color legend, and increased the font size of the model
confidence contents to reduce the amount of whitespace present.