508 compliance changes
Based on the feedback from two tools (HTML CodeSniffer and WAVE),
and the issues already present on GitHub, made changes to the
code to move towards 508 compliance.

This isn't a comprehensive set of changes, but with these the
issues from our autochecking tools that couldn't be safely ignored
are taken care of.