Species Select and Image Bugs
When hovering over 'Clear selection', it would attempt to find an image for that but as it is not a species, that is not needed and an image should not be displayed. Also, the image div is clickable when there is no species selected so the user can open the large image overlay with no image. Both of these have been fixed.